+++ benchmark/benchmark_rank_from_data.cu
@@ -57,7 +57,7 @@ static void BM_Rank(T const* data, size_t const data_size,
         size_t const num_iters =
             std::max<size_t>(1, kBenchmarkTime / (warmup_time / 5));
         times.resize(num_iters);
-        for (int i = 0; i < num_iters; ++i) {
+        for (size_t i = 0; i < num_iters; ++i) {
           start = std::chrono::high_resolution_clock::now();
           auto results = wt.rank(queries.data(), query_num);
           end = std::chrono::high_resolution_clock::now();
